(feat) Make embedded editor react to schema changes
