+++ src/components/form-editor/form-editor.scss
@@ -31,20 +31,3 @@
   margin-left: 0;
   margin-right: 0;
 }
-
-.actionButtons {
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  margin: 1rem 0;
-
-  button {
-    margin-left: 1rem
-  }
-}
-
-.spinner {
-  &:global(.cds--inline-loading) {
-    min-height: 1rem;
-  }
-}
