+++ src/components/form-renderer/form-renderer.component.tsx
@@ -1,16 +1,19 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
+import { InlineLoading, Tile } from "@carbon/react";
 import { OHRIFormSchema, OHRIForm } from "@ohri/openmrs-ohri-form-engine-lib";
-import { Tile } from "@carbon/react";
 import { useConfig } from "@openmrs/esm-framework";
+
+import ActionButtons from "../action-buttons/action-buttons.component";
 import styles from "./form-renderer.scss";
 
 type FormRendererProps = {
+  isLoading: boolean;
   onSchemaChange?: (schema: OHRIFormSchema) => void;
   schema: OHRIFormSchema;
 };
 
-const FormRenderer: React.FC<FormRendererProps> = ({ schema }) => {
+const FormRenderer: React.FC<FormRendererProps> = ({ isLoading, schema }) => {
   const { t } = useTranslation();
   const { patientUuid } = useConfig();
 
@@ -53,29 +56,44 @@ const FormRenderer: React.FC<FormRendererProps> = ({ schema }) => {
     }
   }, [schema]);
 
-  return (
-    <div className={styles.container}>
-      {!schema && (
-        <Tile className={styles.emptyStateTile}>
-          <h4 className={styles.heading}>
-            {t("noSchemaLoaded", "No schema loaded")}
-          </h4>
-          <p className={styles.helperText}>
-            {t(
-              "formRendererHelperText",
-              "Load a form schema in the Schema Editor to the left to see it rendered here by the Form Engine."
-            )}
-          </p>
-        </Tile>
-      )}
-      {schema === schemaToRender && (
-        <OHRIForm
-          formJson={schemaToRender}
-          mode={"enter"}
-          patientUUID={patientUuid}
+  if (isLoading) {
+    return (
+      <div className={styles.loadingContainer}>
+        <InlineLoading
+          className={styles.loader}
+          description={t("loading", "Loading") + "..."}
         />
-      )}
-    </div>
+      </div>
+    );
+  }
+
+  return (
+    <>
+      <ActionButtons schema={schema} t={t} />
+
+      <div className={styles.container}>
+        {!schema && (
+          <Tile className={styles.emptyStateTile}>
+            <h4 className={styles.heading}>
+              {t("noSchemaLoaded", "No schema loaded")}
+            </h4>
+            <p className={styles.helperText}>
+              {t(
+                "formRendererHelperText",
+                "Load a form schema in the Schema Editor to the left to see it rendered here by the Form Engine."
+              )}
+            </p>
+          </Tile>
+        )}
+        {schema === schemaToRender && (
+          <OHRIForm
+            formJson={schemaToRender}
+            mode={"enter"}
+            patientUUID={patientUuid}
+          />
+        )}
+      </div>
+    </>
   );
 };
