+++ src/components/interactive-builder/interactive-builder.scss
@@ -54,12 +54,22 @@
 }
 
 .header {
-  margin-top: 5rem; 
-  margin-bottom: 1rem;
-  width: 80%;
+  margin: 2rem 0;
 }
 
 .explainer {
+  margin-bottom: 2rem;
+
+  p {
+    margin: 1.5rem 0;
+
+    &:first-of-type {
+      @include type.type-style('heading-04');
+    }
+  }
+}
+
+.sectionExplainer {
   margin: 1.5rem 0;
 }
