+++ translations/en.json
@@ -3,6 +3,7 @@
   "addQuestion": "Add Question",
   "addSection": "Add Section",
   "cancel": "Cancel",
+  "cancelButtonText": "Cancel",
   "checkFilters": "Check the filters above",
   "chooseEncounterType": "Choose an encounter type to link your form to",
   "chooseFieldType": "Choose a field type",
@@ -16,6 +17,8 @@
   "description": "Description",
   "descriptionPlaceholderText": "e.g. A form used to collect encounter data for clients in the Express Care program.",
   "downloadSchema": "Download schema",
+  "editNameButton": "Edit",
+  "editQuestion": "Edit question",
   "editSchema": "Edit schema",
   "encounterType": "Encounter Type",
   "enterConceptName": "Enter concept name",
@@ -28,8 +31,12 @@
   "errorCreatingQuestion": "Error creating question",
   "errorCreatingSection": "Error creating section",
   "errorPublishingForm": "Error publishing form",
+  "errorRenamingForm": "Error renaming form",
+  "errorRenamingPage": "Error renaming page",
+  "errorRenamingSection": "Error renaming section",
   "errorUnpublishingForm": "Error unpublishing form",
   "errorUpdatingForm": "Error updating form",
+  "errorUpdatingQuestion": "Error updating question",
   "expandSectionExplainer": "Below are the sections linked to this page. Expand each section to add questions to it.",
   "fieldType": "Field type",
   "filterByPublishedStatus": "Filter by publish status",
@@ -39,6 +46,7 @@
   "formName": "Form name",
   "formPublished": "Form published",
   "formPublishedSuccessfully": "form was published successfully",
+  "formRenamed": "Form renamed",
   "formRendererHelperText": "Load a form schema in the Schema Editor to the left to see it rendered here by the Form Engine.",
   "forms": "Forms",
   "formUnpublished": "Form unpublished",
@@ -49,6 +57,7 @@
   "interactiveBuilderHelperText": "The Interactive Builder lets you build your form schema without writing JSON code. The Preview on the right automatically updates as you build your form. When done, click Save Form to save your form.",
   "isQuestionRequiredOrOptional": "Is this question a required or optional field? Required fields must be answered before the form can be submitted.",
   "labelPlaceholder": "e.g. Type of Anaesthesia",
+  "loading": "Loading",
   "loadingSchema": "Loading schema",
   "name": "Name",
   "no": "No",
@@ -59,6 +68,7 @@
   "optional": "Optional",
   "pageCreated": "New page created",
   "pageExplainer": "Pages typically have one or more sections. Click the button below to add a section to your page.",
+  "pageRenamed": "Page renamed",
   "preview": "Preview",
   "published": "Published",
   "publishForm": "Publish Form",
@@ -77,6 +87,7 @@
   "save": "Save",
   "saveAsModal": "Would you want to update the form or save as a new version?",
   "saveAsNewForm": "Save as a new version",
+  "saveButtonText": "Save",
   "saveConfirmation": "Save or Update form",
   "saveExplainerText": "Clicking the Save button saves your form schema to the database. To see your form in your frontend, you first need to publish it. Click the Publish button to publish your form.",
   "saveForm": "Save Form",
@@ -93,6 +104,7 @@
   "searchThisList": "Search this list",
   "sectionCreated": "New section created",
   "sectionExplainer": "A section will typically contain one or more questions. Click the button below to add a question to this section.",
+  "sectionRenamed": "Section renamed",
   "selectAnswersToDisplay": "Select answers to display",
   "selectBackingConcept": "Select backing concept",
   "startBuilding": "Start building",
@@ -105,6 +117,7 @@
   "updateExistingForm": "Update existing version",
   "validFieldTypeRequired": "A valid field type value is required",
   "version": "Version",
-  "welcomeText": "Welcome to the Interactive Schema builder! Add pages, sections and questions to your form. The Preview on the right automatically updates as you build your form. When done, click Save Form to save your form.",
+  "welcomeExplainer": "Add pages, sections and questions to your form. The Preview tab automatically updates as you build your form. When finished, click Save Form to save your form.",
+  "welcomeHeading": "Interactive schema builder",
   "yes": "Yes"
 }
